Adapt remote branch listing test to CLI git 2.48.0 (#1242)
* Adapt GitAPITest for CLI git 2.48.0

Fetch from CLI git 2.48.0 and later return origin/HEAD

https://github.blog/open-source/git/highlights-from-git-2-48/ says:

> With Git 2.48, if the remote has a default branch but refs/remotes/origin/HEAD
> is missing locally, then a fetch will update it.

This test was unintentionally testing the behavior of CLI git before
2.48.0.  Other tests in GitClientFetchTest were testing that origin/HEAD
was reported as a branch.

* Show names/sha1 of detected branches on failure

When the assertion fails, show the details of the set of branches so
that the test failure has more information about the failure conditions.

Thanks to @darinpope for the question that prompted the improvement

* Remove unused cliGitCommand from GitAPITest

(cherry picked from commit 8d5f9c37c25ece914f7d968efda9bdf36ad74aab)