Skip ssh host tests on ci.jenkins.io
The GitHub.com SSH abuse prevention system rejects connections from
the tests when they are run on ci.jenkins.io.  These tests should only
be run as end to end acceptance tests but they are not too disruptive
(for now) to continue to be included in the default tests.

Additional tests should be created that use test containers so that they
can use the ssh connection on the test container instead of using the
GitHub.com SSH servers.

Unfortunately, I can't prioritize that test rewrite now, so these tests
are disabled on ci.jenkins.io.