Use try with resources in JGit doCheckoutWithResetAndRetry
The JGit implementation for doCheckoutWithResetAndRetry needs to
remove conflicting files that might have been created during a
checkout.  The conflicting file removal was previously done by closing
the repository and opening the repository again for the cleanup.

Rather than use such complicated logic with the risk that files won't
be closed on exceptions, let's switch to try with resources so that
Java will close the repository on exit from the block.