Ignore assertion for ed_25519 key test
Apache Mina SSHD prior to 2.15.0 does not support ed_25519 keys.
The assertion in that case was asserting that host keys were not checked.
Apache Mina SSHD 2.15.0 supports ed_25519 keys.  The assertion in that
case should correctly assert that the host keys are checked.

9c1e75f7e8504c7f23eed4e049bb7e4e4c87d0ec on the master branch correctly
depends on Apache Mina SSHD 2.15.0 and checks the ed_25519 host key.

This ignores the assertion so that plugin BOM tests will not fail when
Apache Mina SSHD 2.15.0 is used.