[CVE-2025-4049] XXE vulnerability in Eclipse JGit (non-impacting)
https://projects.eclipse.org/projects/technology.jgit/releases/7.0.1
reports that CVE-2025-4049 is fixed in JGit 7.0.1.

https://www.cve.org/CVERecord?id=CVE-2025-4949 describes the issue
as :

> the ManifestParser class used by the repo command and the AmazonS3
> class used to implement the experimental amazons3 git transport protocol
> allowing to store git pack files in an Amazon S3 bucket, are vulnerable
> to XML External Entity (XXE) attacks when parsing XML files. This
> vulnerability can lead to information disclosure, denial of service,
> and other security issues.

This upgrade is a safeguard against scanner reports of a vulnerability
that does not actually affect the git client plugin.  The vulnerability
does not apply to the git client plugin because the git client plugin
does not use the ManifestParser class or the RepoCommand class that
depends on ManifestParser.

A GitHub search showed no use of RepoCommand in any of the jenkinsci
organization.