Explicitly declare bundled artifacts
Do not allow additions to the bundled dependencies without a change to
the list of bundledArtifacts.  Avoid the mistakes made in the past when
a new transitive dependency was inadvertently added to the plugin and
bundled into the hpi / jpi file.

Draft until the release of pull request:

* https://github.com/jenkinsci/maven-hpi-plugin/pull/771

Testing done:

Confirmed that compilation passes and a single test passes.  Rely on
ci.jenkins.io for more complete testing