[JENKINS-76004] Use git log instead of git whatchanged for changelog (#1326)
Command line git 2.51.0 has deprecated the "git whatchanged" command and
recommends that it be replaced with "git log --raw".  Refer to the blog
post: https://github.blog/open-source/git/highlights-from-git-2-51/

Testing done:

Interactive testing confirmed that without this change, the changelog
that is usually displayed on a freestyle project will silently remain
empty when using command line git 2.51.0.

With this change, the changelog displays as expected.

Automated tests pass on:

* Debian 11 (Git 2.30)
* Debian 12 (Git 2.39)
* FreeBSD 14 (Git 2.50)
* Red Hat 8 (Git 2.51, built from source code)
* Ubuntu 22 (Git 2.34)
* Ubuntu 24 (Git 2.43)
* Windows 11 (Git 2.50.1)

Those test combinations are a wide range of versions of command line git.