Improve JMH sample repository choices
Include a wider range of small repositories in order to be more likely
to detect shifts in the turning point where command line git performs
better than JGit.

Current data shows that JGit is faster than command line git for
repositories up to 8 MB.  Command line git is dramatically faster than
JGit for the 450 MB repository that is the largest repository used in
the test.