Run GitClientTest.testSparseCheckoutWithCliGitLFS when credentials available (#1337)
The test fails in nearly 100% of cases when credentials are not provided,
even though the repository is a public repository.  When credentials are
provided, the test only fails in about 20% of the cases.  The credentials
I use for the test are narrowly scoped to exactly one repository and to
only allow access to repository content.  Since the repository content is
already publicly readable, the credential does not grant any additional
permissions.

The ci.jenkins.io agents do not have the credential, so they will not
run the test.

Testing done

Confirmed that test fails consistently without credentials.

Confirmed that test fails in approximately 20% of cases when I use a
range of agents including:

* Debian 11
* Debian 12
* Debian Testing
* openSUSE 15.6
* Red Hat Enterprise Linux 8
* Rocky Linux 9
* Ubuntu 22
* Ubuntu 24
* Windows