Use system temporary directory for known_hosts file (#1709)
Windows ssh.exe 9.5p1 and ssh.exe 9.5p2 will not read the known_hosts
file if there is a space character in the path to the file.

Since we already use the system temporary directory for other scripts that
do not contain sensitive information, we'll use it for the known_hosts
file as well.  The contents of the known_hosts file is not sensitive
information.

Testing done:

* Confirmed that without this change, Windows ssh.exe does not read the
  known_hosts file when using a manually verified known_hosts file from
  a workspace that has spaces in its absolute path.

* Confirmed that with this change, Windows ssh.exe reads the known_hosts
  file when using a manually verified known_hosts file from a workspace
  that has spaces in its absolute path.

Fixes #1703