Do not use paths with '%' for SSH temporary files (#1712)
SSH has special rules related to '%' characters in its file name
arguments.  Refer to the ssh_config man page for its descriptions
of tokens:

https://man.openbsd.org/ssh_config#TOKENS

If the absolute path to the temporary file for the private key includes a
'%' character, then use the system temporary directory instead.

Likewise if the absolute path to the Windows temporary file for
the private key includes a ' ', a '(', or a ')', then use the system
temporary directory.

Makes the createTempFile() implementation match the comment.

Testing done:

* Confirmed that I can duplicate the bug with a multibranch
  Pipeline cloned through an SSH URL when one of the branches in
  the repository contains a '/' character.  In my test case, I used
  'bug/git-cleint-plugin-1710' as the branch name
* Confirmed that the bug is resolved by this change

Pipeline definition used for test:

pipeline {
    agent {
        label 'windows'
    }
    options {
        skipDefaultCheckout true
    }
    stages {
        stage('Checkout') {
            steps {
                ws('percent%2Fencoded') {
                    checkout scm
                }
                ws('a b') {
                    checkout scm
                }
                ws('c(') {
                    checkout scm
                }
                ws('d)') {
                    checkout scm
                }
                ws('e())') {
                    checkout scm
                }
                ws('f g') {
                    checkout scm
                }
            }
        }
    }
}